ImageEdgeDetection.jl Documentation

A Julia package containing a number of algorithms for detecting edges in images.

ImageEdgeDetection.jl Documentation

Getting started

Basic usage

Getting started

This package is part of a wider Julia-based image processing ecosystem. If you are starting out, then you may benefit from reading about some fundamental conventions that the ecosystem utilizes that are markedly different from how images are typically represented in OpenCV, MATLAB, ImageJ or Python.

The usage examples in the ImageEdgeDetection.jl package assume that you have already installed some key packages. Notably, the examples assume that you are able to load and display an image. Loading an image is facilitated through the FileIO.jl package, which uses QuartzImageIO.jl if you are on MacOS, and ImageMagick.jl otherwise. Depending on your particular system configuration, you might encounter problems installing the image loading packages, in which case you can refer to the troubleshooting guide.

Image display is typically handled by the ImageView.jl package. Alternatives include the various plotting packages, including Makie.jl. There is also the ImageShow.jl package which facilitates displaying images in Jupyter notebooks via IJulia.jl. Finally, one can also obtain a useful preview of an image in the REPL using the ImageInTerminal.jl package. However, this package assumes that the terminal uses a monospace font, and tends not to produce adequate results in a Windows environment.

Another package that is used to illustrate the functionality in ImageEdgeDetection.jl is the TestImages.jl which serves as a repository of many standard image processing test images.

Basic usage

TODO

ImageEdgeDetection.Canny

ImageEdgeDetection.NonmaximaSuppression

ImageEdgeDetection.Percentile

ImageEdgeDetection.suppress_non_maxima!

ImageEdgeDetection.trace_and_threshold!

ImageEdgeDetection.Canny — Type

    Canny <: AbstractEdgeDetectionAlgorithm
    Canny(; spatial_scale = 1, high = Percentile(80), low = Percentile(20))

    detect_edges([T,] img, f::Canny)
    detect_edges!([out,] img, f::Canny)

Returns a binary image depicting the edges of the input image.

Details

TODO

Options

Various options for the parameters of the detect_edges function and Canny type are described in more detail below.

Choices for img

The detect_edges function can handle a variety of input types. By default the type of the returned image is Gray{Bool}.

For colored images, the input is converted grayscale.

Choices for spatial_scale in Canny.

The spatial_scale determines the radius (σ) of the Gaussian filter. It must be a positive real number.

Choices for high and low in Canny.

The hysteresis thresholds high and low (high > low) can be specified as positive numbers, or as Percentiles. If left unspecified, a default value of high = Percentile(80) and low = Percentile(20) is assumed.

Example

using TestImages, FileIO, ImageView

img =  testimage("mandril_gray")
img_edges = detect_edges(img, Canny(spatial_scale = 1.4))

imshow(img)
imshow(img_edges)

References

J. Canny, "A Computational Approach to Edge Detection," in IEEE Transactions on Pattern Analysis and Machine Intelligence, vol. PAMI-8, no. 6, pp. 679-698, Nov. 1986, doi: 10.1109/TPAMI.1986.4767851.

ImageEdgeDetection.NonmaximaSuppression — Type

    NonmaximaSuppression <: AbstractEdgeThinningAlgorithm
    NonmaximaSuppression(; threshold::Union{Number, Percentile} = Percentile(20))

    f = NonmaximaSuppression()
    f(out::AbstractArray, mag::AbstractArray, g₁::AbstractArray, g₂::AbstractArray, f::NonmaximaSuppression)

Isolates local maxima of gradient magnitude mag along the local gradient direction and stores the result in out. The arguments g₁ and g₂ represent the gradient in the first spatial dimension (y), and the second spatial dimension (x), respectively.

Details

TODO

Example

using TestImages, FileIO, ImageView

img =  testimage("mandril_gray")

# Gradient in the first and second spatial dimension
g₁, g₂ = imgradients(img, KernelFactors.scharr)

# Gradient magnitude
mag = hypot.(g₁, g₂)

nms = zeros(eltype(mag), axes(mag))
# Instantiate the NonmaximaSuppression functor.
f = NonmaximaSuppression()

# Suppress the non-maximal gradient magnitudes and store the result in `nms`.
f(nms, mag, g₁, g₂)

imshow(img)
imshow(mag)
imshow(nms)

References

J. Canny, "A Computational Approach to Edge Detection," in IEEE Transactions on Pattern Analysis and Machine Intelligence, vol. PAMI-8, no. 6, pp. 679-698, Nov. 1986, doi: 10.1109/TPAMI.1986.4767851.

ImageEdgeDetection.Percentile — Type

Percentile(x)

Indicate that x should be interpreted as a percentile rather than an absolute value. For example,

detect_edges(img, Canny(high = 80, low = 20)) uses absolute thresholds on the edge magnitudes

detect_edges(img, Canny(high = Percentile(80), low = Percentile(20))) uses percentiles of the edge magnitude image as threshold

ImageEdgeDetection.suppress_non_maxima! — Method

    suppress_non_maxima!(out::AbstractArray, mag::AbstractArray, gx::AbstractArray, gy::AbstractArray, threshold::Number)

Isolates local maxima of gradient magnitude by “non-maximum suppression” along the local gradient direction.

ImageEdgeDetection.trace_and_threshold! — Method

    trace_and_threshold!(out::AbstractArray,  mag::AbstractArray, i₀::CartesianIndex, low::Number)

Marks all pixels of an edge that are 8-connected to i₀ and exhibit a gradient magnitude above low.