Function References

Function References

General function

Edge Detection Algorithms

Canny

Edge Thinning Algorithms

Non-maxima Suppression

Supplementary Types

General function

ImageEdgeDetection.EdgeDetectionAPI.detect_edges — Function

detect_edges([T::Type,] img, f::AbstractEdgeDetectionAlgorithm, args...; kwargs...)

Detect edges of img using algorithm f.

Output

The return image img_edges is an Array{T}.

If T is not specified, then it's inferred.

Examples

Just simply pass the input image and algorithm to detect_edges

img_edges = detect_edges(img, f)

This reads as "detect_edges of image img using algorithm f".

You can also explicitly specify the return type:

img_edges_float32 = detect_edges(Gray{Float32}, img, f)

See also detect_edges! for in-place edge detection.

ImageEdgeDetection.EdgeDetectionAPI.detect_edges! — Function

detect_edges!([out,] img, f::AbstractEdgeDetectionAlgorithm, args...; kwargs...)

Detect edges of img using algorithm f.

Output

If out is specified, it will be changed in place. Otherwise img will be changed in place.

Examples

Just simply pass an algorithm to detect_edges!:

img_edges = similar(img)
detect_edges!(img_edges, img, f)

For cases you just want to change img in place, you don't necessarily need to manually allocate img_edges; just use the convenient method:

detect_edges!(img, f)

See also: detect_edges

ImageEdgeDetection.EdgeDetectionAPI.thin_edges — Function

thin_edges([T::Type,] mag, g₁, g₂, f::AbstractEdgeThinningAlgorithm, args...; kwargs...)

Using algorithm f, thin the edge-response based on the edge magnitude mag, the gradient in the first spatial dimension g₁, and the gradient in the second spatial dimension g₂.

Output

Returns an Array{T} representing the thinned edge response.

If T is not specified, then it's inferred.

Examples

Just simply pass the input image and algorithm to thin_edges

using TestImages, ImageFiltering
img =  Gray.(testimage("mandril"))

# Gradient in the first and second spatial dimension
g₁, g₂ = imgradients(img, KernelFactors.scharr)

# Gradient magnitude
mag = hypot.(g₁, g₂)

thinned_edges = thin_edges(mag, g₁, g₂, f)

This reads as "thin_edges based on the edge response magnitude, and spatial gradients, using algorithm f".

You can also explicitly specify the return type:

thinned_edges_float32 = thin_edges(Gray{Float32}, mag, g₁, g₂, f)

See also thin_edges! for in-place edge thinning.

ImageEdgeDetection.EdgeDetectionAPI.thin_edges! — Function

thin_edges!([out,] mag, g₁, g₂, f::AbstractEdgeThinningAlgorithm, args...; kwargs...)

Isolate local maxima of gradient magnitude mag along the local gradient direction. The arguments g₁ and g₂ represent the gradient in the first spatial dimension (y), and the second spatial dimension (x), respectively.

Output

If out is specified, it will be changed in place. Otherwise mag will be changed in place.

Examples

Just simply pass an algorithm to thin_edges!:

using TestImages, ImageFiltering
img =  Gray.(testimage("mandril"))

# Gradient in the first and second spatial dimension
g₁, g₂ = imgradients(img, KernelFactors.scharr)

# Gradient magnitude
mag = hypot.(g₁, g₂)

thinned_edges = zeros(eltype(mag), axes(mag))
thin_edges!(thinned_edges, mag, g₁, g₂, f)

For cases you just want to change mag in place, you don't necessarily need to manually allocate thinned_edges; just use the convenient method:

thin_edges!(mag, g₁, g₂, f)

See also: thin_edges

Edge Detection Algorithms

ImageEdgeDetection.EdgeDetectionAPI.AbstractEdgeDetectionAlgorithm — Type

AbstractEdgeDetectionAlgorithm <: AbstractImageFilter

The root type for ImageEdgeDetection package.

Any concrete edge detection algorithm shall subtype it to support detect_edges and detect_edges! APIs.

Examples

All edge detection algorithms in ImageEdgeDetection are called in the following pattern:

# first generate an algorithm instance
f = Canny()

# then pass the algorithm to `detect_edges`
img_edges, list_edges = detect_edges(img, f)

# or use in-place version `detect_edges!`
img_edges = similar(img)
list_edges = detect_edges!(img_edges, img, f)

For more examples, please check detect_edges, detect_edges! and concrete algorithms.

Canny

ImageEdgeDetection.Canny — Type

    Canny <: AbstractEdgeDetectionAlgorithm
    Canny(; spatial_scale = 1, high = Percentile(80), low = Percentile(20))

    detect_edges([T,] img, f::Canny)
    detect_edges!([out,] img, f::Canny)

Returns a binary image depicting the edges of the input image.

Details

TODO

Options

Various options for the parameters of the detect_edges function and Canny type are described in more detail below.

Choices for img

The detect_edges function can handle a variety of input types. By default the type of the returned image matches the type of the input image.

For colored images, the input is converted grayscale.

Choices for spatial_scale in Canny.

The spatial_scale determines the radius (σ) of the Gaussian filter. It must be a positive real number.

Choices for high and low in Canny.

The hysteresis thresholds high and low (high > low) can be specified as positive numbers, or as Percentiles. If left unspecified, a default value of high = Percentile(80) and low = Percentile(20) is assumed.

Example

using TestImages, FileIO, ImageView

img =  testimage("mandril_gray")
img_edges = detect_edges(img, Canny(spatial_scale = 1.4))

imshow(img)
imshow(img_edges)

References

J. Canny, "A Computational Approach to Edge Detection," in IEEE Transactions on Pattern Analysis and Machine Intelligence, vol. PAMI-8, no. 6, pp. 679-698, Nov. 1986, doi: 10.1109/TPAMI.1986.4767851.

Edge Thinning Algorithms

ImageEdgeDetection.EdgeDetectionAPI.AbstractEdgeThinningAlgorithm — Type

AbstractEdgeThinningAlgorithm <: AbstractImageFilter

A root type for ImageEdgeDetection package.

Any concrete edge thinning algorithm shall subtype it to support thin_edges and thin_edges! APIs.

Examples

All edge thinning algorithms in ImageEdgeDetection are called in the following pattern:

# first generate an algorithm instance
f = NonmaximaSuppression()

# determine the image gradients
g₁, g₂ = imgradients(img, KernelFactors.scharr)

# determine the gradient magnitude
mag = hypot.(g₁, g₂)

# then pass the algorithm to `thin_edges`
thinned_edges = thin_edges(mag, g₁, g₂, f)

# or use in-place version `thin_edges!`
thinned_edges = zeros(eltype(mag), axes(mag))
thin_edges!(thinned_edges, mag, g₁, g₂, f)

For more examples, please check thin_edges, thin_edges! and concrete algorithms.

Non-maxima Suppression

ImageEdgeDetection.NonmaximaSuppression — Type

    NonmaximaSuppression <: AbstractEdgeThinningAlgorithm
    NonmaximaSuppression(; threshold::Union{Number, Percentile} = Percentile(20))

    f = NonmaximaSuppression()
    f(out::AbstractArray, mag::AbstractArray, g₁::AbstractArray, g₂::AbstractArray, f::NonmaximaSuppression)

Isolates local maxima of gradient magnitude mag along the local gradient direction and stores the result in out. The arguments g₁ and g₂ represent the gradient in the first spatial dimension (y), and the second spatial dimension (x), respectively.

Details

TODO

Example

using TestImages, FileIO, ImageView, ImageEdgeDetection, ImageFiltering

img =  testimage("mandril_gray")

# Gradient in the first and second spatial dimension
g₁, g₂ = imgradients(img, KernelFactors.scharr)

# Gradient magnitude
mag = hypot.(g₁, g₂)

nms = zeros(eltype(mag), axes(mag))
# Instantiate the NonmaximaSuppression functor.
f = NonmaximaSuppression()

# Suppress the non-maximal gradient magnitudes and store the result in `nms`.
f(nms, mag, g₁, g₂)

imshow(img)
imshow(mag)
imshow(nms)

References

J. Canny, "A Computational Approach to Edge Detection," in IEEE Transactions on Pattern Analysis and Machine Intelligence, vol. PAMI-8, no. 6, pp. 679-698, Nov. 1986, doi: 10.1109/TPAMI.1986.4767851.

Supplementary Types

ImageEdgeDetection.Percentile — Type

Percentile(x)

Indicate that x should be interpreted as a percentile rather than an absolute value. For example,

detect_edges(img, Canny(high = 80, low = 20)) uses absolute thresholds on the edge magnitudes

detect_edges(img, Canny(high = Percentile(80), low = Percentile(20))) uses percentiles of the edge magnitude image as threshold

ImageEdgeDetection.Canny

ImageEdgeDetection.EdgeDetectionAPI.AbstractEdgeDetectionAlgorithm

ImageEdgeDetection.EdgeDetectionAPI.AbstractEdgeThinningAlgorithm

ImageEdgeDetection.NonmaximaSuppression

ImageEdgeDetection.Percentile

ImageEdgeDetection.EdgeDetectionAPI.detect_edges

ImageEdgeDetection.EdgeDetectionAPI.detect_edges!

ImageEdgeDetection.EdgeDetectionAPI.thin_edges

ImageEdgeDetection.EdgeDetectionAPI.thin_edges!